Change chess piece height
Because the chessboard grid is set to
a specific height (60px), changing the 
height of the chess piece images
to 50px will achieve a better look.  
Along with that changing the top and 
bottom margin to 5px each will create
a better looking image, especially if 
the page is resized.