Auto center board, add max-width
Originally, the board scaled with 
increased horizontal stretching, but
it creates a contorted image. So instead
of having it continuously stretch, we added
a max-width to stop it from stretching any
larger than a 1:1 ratio with the height. This
creates a much cleaner look. Adding the auto
centering makes sure that the board stays
centered.