Clarification
There may be confusion as to the
commits of this assignment, since this
is a personal assignment and my group
partner (for assignment a2) has committed
to this file.  All of the work in this version
is my own, and I did not receive help to
work on this.  Aatiqa accidently pushed
their version onto my repository thinking
it was theirs.