Add a container div inside body
There was an issue with overflow being renderd outside of the browser
if the body's height is set to 100% and then also flexed.  To fix this
add a container div that is flexed, and unflex the body.  The body will
take up 100% of the viewport, and the flex container will allow for
overflow to be properly rendered.